#287 - Move description list to shared components folder
